LayoutTests/imported/w3c/web-platform-tests/html/dom/idlharness-shadowrealm.window-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: idl_test_shadowrealm

Harness Error (FAIL), message = ReferenceError: Can't find variable: idl_test_shadowrealm
